refactor: Move SchemaModule to modules dir
